requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask-sock==0.7.0
gunicorn==22.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
python-engineio==4.9.1
python-socketio==5.11.2
simple-websocket==1.0.0
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
